Introduce initial version of rline crate
This change introduces the first version of the rline crate. The crate
is a wrapper around libreadline's "Alternate Interface". It provides the
goodness of libreadline (its powerfulness for inputting text as well as
its configurability) while leaving the actual character input in the
hands of a user. This can be a powerful tool for terminal based (but not
necessarily command line based) applications that want to handle input
on their own terms and only harness libreadline for a subset of their
text input needs. But even graphical applications which typically are
not using file stream based input (or a terminal for that matter) at all
are enabled to use libreadline for providing an input experience as
usually only known on the command line.
In a nutshell, by instantiating a Readline object as provided by this
crate one gets the ability to feed individual characters into a buffer
and libreadline handles buffer management, higher level cursor movements
(e.g., skipping over words etc.), undo functionality, and so on.