Add support for "inspecting" the current line state
While in some cases it is sufficient to feed characters to the crate and
only get informed about completed lines, in many cases it is actually
important to also be able to inspect the current line state (that is,
the text as well as the current cursor position).
With this change we add a method that allows for precisely such an
access to happen. The API is flexible in that allows for allocation free
inspection while also providing the means for copying out the current
state.