Add support for resetting the current line state
Sometimes it is necessary to not start editing an empty line, but rather
start off with some predefined content. Currently, working with such a
use case is not possible.
With this change we introduce a new method, reset, that can be used to
reset the current line state (i.e., text and cursor position) to the one
supplied. In addition to this data, callers need to specify whether or
not to clear libreadline's undo buffer for this line.