Free libreadline allocated memory to avoid one-time memory leak
Because we allow for different libreadline contexts to be active, each
representing it at a different state, we need to manage a per-context
memory buffers. When we initially create the first such context, we
effectively overwrite the existing libreadline state and with it the
buffers already allocated. That means this buffer memory is effectively
lost.
With this change we make sure to free this memory, in order to prevent
this one-time memory leak.