Install line handler callback only once
This change fixes a problem that manifested as the initial configuration
no longer being effective when multiple Readline contexts are in use.
The reason for this problem has to do with libreadline internals. When
we first call rl_callback_handler_install the library is initialized. As
part of this initialization we read in the user-configuration.
Unfortunately, as this initialization happens only once, all other
contexts will not have it, as they will start off with a state from
earlier. Hence, only the very first context would have the
user-configuration active.
So with this change we move said call into the one-time state
initialization function. A test is added to prevent regressions from
happening. Note that said test is implemented as an integration test.
That is a requirement because that will cause it to run without
interference from other threads. That in turn is necessary because we
set global configuration without any guarantee for mutual exclusion in a
multi-threaded environment.