Clarify that Readline::reset does not reset input mode
In rare (but possible) scenarios, a user of this crate may want to
manually cancel input through libreadline prematurely, i.e., before it
actually completed a line. Because libreadline effectively does not know
about this cancellation (as it can and will happen completely outside of
its control, by virtue of no longer funneling characters to it), it will
stay in the same mode it was in before the cancellation. That can be a
problem, because that mode will in turn still be active when input is
resumed, where typically the assumption is that "normal" input mode is
active.
That is a known limitation and no sufficiently suitable solution has
been found to date. With this change we add a note about this behavior
to the Readline::reset method.