Fix wrong lifetime being used in Readline::inspect method
The function being invoked in Readline's inspect method receives a
reference to a CStr object. However, the lifetime of said reference is
too broad and would even allow for returning of the reference itself.
That would be wrong because the CStr actually references libreadline's
rl_line_buffer variable and that could be changed transparently after
the method returned (because the lock is dropped).
To fix this problem this change corrects the lifetime of said reference,
preventing a direct return.