Remove #![deny(warnings)] attribute
We have been burned in the past by having a 'deny(warnings)' attribute
in another crate that caused compile errors because a function we used
got deprecated. This is a potentially huge problem and must be avoided
at all cost.
With this change we demote lints that are suspected to be potentially
prone to incompatible changes of this sort from a 'deny' to a 'warn'.
Similarly, we remove the 'warnings' lint entirely, as those warnings are
already emitted by default.