Add example illustrating how to use the Readline objects
In order to hit the ground running when using a new library it is often
helpful if a boiled down example provides some working code to start off
with.
This change adds such an example. The example works on the command line,
but the basics are roughly transferable to GUI applications as well. It
uses the termion crate to transition the terminal into raw mode and then
reads individual characters and feeds them to a `Readline` object, which
will process them and report the final result, which in turn is printed
to stdout.
Note that not all escape sequences are handled properly. The cursors
keys, for example, trigger unexpected behavior. This problems is due to
current limitations in our API which do not allow user code to pass in
multi-byte byte sequences.