Fix use-after-free bug in user-configuration test
This change fixes a use-after-free bug in the user-configuration test.
The problem was that we converted a CString into a Vec<u8> and then
interpreted it directly as a pointer. Because we lack a let binding for
the resulting intermediate variable, the vector will be destroyed pretty
much immediately and we are working on a dangling pointer.