Adjust Readline::feed method to accept a byte instead of an int
libreadline, despite accepting an int to the rl_stuff_char function,
actually just treats the value passed in as a char, i.e., a single byte.
This can cause confusion when it comes to Unicode characters. To make
this problem more explicit, this change adjusts the signature of the
Readline::feed method to accept a u8 instead of the previously used
c_int.