Adjust Readline::feed to accept a byte-sequence
With the recent addition of an example illustrating the usage of the
crate, we have found that the Readline::feed method is insufficient for
advanced input purposes. In particular, certain key presses are
represented as multi-byte sequences and while feeding them to
libreadline one byte at a time is a feasible approach (and, in fact,
necessary for lack of a more suitable API provided by libreadline), we
must not ask the library to process the input after each byte (through
the usage of rl_callback_read_char). Doing so can lead to escape
sequences being broken up and, with that, result in different key
presses being recognized.
To fix this problem, this change adjusts the feed method to accept a
byte slice instead.
Note that we have chosen to accept a byte slice instead of an iterator
over bytes, despite the latter being the more generalized concept,
because ultimately it is considered very likely that the user already
has either a vector or a slice of bytes by virtue of key presses pretty
much always being read through an API that works with such a buffer. It
is also more cumbersome to work with iterators as that requires an
additional conversion step at each invocation of the method in the user
code.