Use cargo:rustc-link-lib instead of manually setting rustc-flaggs
The more appropriate way of linking against libraries from within a
build script in Cargo is by emitting the cargo:rustc-link-lib specifier.
cargo:rustc-flags accomplished the same goal, but it is unclear whether
that may cause problems down the line because Cargo cannot understand
this directive in the same way as it can rustc-link-lib.
Hence, with this change we adjust the build.rs file to emit the more
appropriate output.