Revamp README file and add more usage details
The README has always been considered a bit sporadic. However, having a
well-written README presenting the crate in an easy-to-understand manner
and clarifying typical questions is very important for leaving a good
impression with potential users.
To that end, this change revamps the existing README, reformulating &
reformatting some parts while adding a brief overview on what typical
usage of the crate looks like.