Directly store readline_state inside of ReadlineGuard
The ReadlineGuard currently stores a reference to the full Readline
object. While that is not wrong, it is more than we actually need. The
only reason we need this reference is to restore the libreadline state
when we drop the guard. Hence, storing a reference to readline_state is
enough.
This change adjusts the guard's implementation accordingly.