Use RefCell for storing readline_state inside Readline
The API surface of the Readline object has one glaring problem: the peek
method, which is supposed to be only reading state, not modifying it,
requires a mutable self to work on.
The problem really comes from the fact that we need to store back
potentially changed state from libreadline after we are done calling
into it, which, for obvious reasons, needs a mutable object. However,
because all of that is an implementation detail, requiring mutability
at the API level does not seem like the right decision.
With this change we resolve this problem with interior mutability by
storing a Readline's readline_state in a std::cell::RefCell.