Remove usage of deprecated std::mem::uninitialized
In order to not have an unnecessary zero initialization in the program,
we decided to use std::mem::uninitialized. This function has since been
deprecated and usage discouraged, as it is hard to use correctly. With
this change we switch over to using std::mem::MaybeUninit instead, which
makes the code somewhat nicer as we can get rid of the usage of an
Option.