Correctly handle multi-byte Unicode characters in example
The "basic" example we provide for illustrating the usage of the library
does not work correctly in conjunction with multi-byte Unicode
characters. More specifically, the cursor is set incorrectly for such
multi-bytes characters.
While libreadline recognizes those characters and handles them
correctly, the reported cursor position is the byte offset from the
start of the buffer. When we set the terminal's cursor we therefore need
to translate this position, taking into consideration which grapheme
cluster this position maps to, as each such grapheme cluster takes up a
single cell on the terminal.