Use non-generic inner function pattern
The Readline::feed() and Readline::reset() methods are generic over an
input argument, but ultimately said genericity evaporates, because it is
a means of accepting more input types that all are trivially convertible
into the actual type we work with. Because Rust effectively instantiates
functions for each different generic type, this could be the cause of
code bloat.
With this change we use the non-generic inner function pattern to
minimize the truly generic portion of the code in an attempt to
eliminate unnecessary code bloat.