Free undo lists as part of Readline object destruction
As it turns out, libreadline may allocate additional data as part of its
undo lists and right now we may end up leaking the memory, because we do
not release it as part of object destruction.
With this change we take care of clearing the undo list of the initial
state and then freeing any data allocated by libreadline.