Heap allocate readline_state objects
The readline_state object that reach Readline instance carries around is
fairly sizable, weighing in at 512 bytes each. We end up copying it
around whenever we move such an object, which is work that can easily be
avoided.
With this change we heap allocate the data, shrinking the size of a
Readline object significantly and causing moves to have much less data
to copy.