Keep original readline buffer sizes
When we create a Readline "context" object we heap allocate buffers that
libreadline requires. So far we have specified our own buffer sizes,
independent of libreadline's defaults. But there really is no good
reason not to go with whatever they were using from the get-go.
Thus, with this change we just allocate as many bytes as libreadline
would by default.