Create GitHub release as part of 'publish' workflow
When we create a new release it is good measure to also create a git tag
and create a corresponding GitHub release for increased visibility and
easy of access. So far we have done that manually, but with this change
we adjust the publish workflow to take care of this task for us.