Only include library in MSRV check
When we check the minimum supported Rust version in CI, we build the
entire crate including all examples and tests. However, we don't really
care about tests and examples adhering to the MSRV -- they are only
development artifacts with relevance to the crate itself.
To improve the situation and decouple those things, adjust the test
workflow in such a way that MSRV testing happens only on the library
itself.