Honor system bitness when configuring library search path
As a convenience for the user, we add both /usr/lib/ and /usr/lib64/ to
the library search path. However, given how linking works, more is not
always better. Specifically, now, on a 64 bit system it is possible that
32 bit archives are search for first and picked (and vice versa for 32
bit systems), which is likely to cause linker failures.
Fix this problem by