Integrate user-configuration test into main test suite
Integrate the user-configuration test into the main test suite. So far
it had been a separate integration test, because it needed to run in a
separate process to ensure isolation. With the test-fork crate we have
the means to run individual unit tests in separate processes, which
seems more convenient.